Ricardo/flask_server/requirements.txt
Flask
Flask-Cors
openai
python-dotenv 